Reverted back to our initial planned course, with a small adjustment. We are now using Sql express container image for our db.
configurations in appsettings, csproj,
startup and commanderContext.cs are done,

### two flag areas to check when debugging:
1. connection string- is timeout correct?
2. startup. automigrate method-
is it supposed to take our db context class
as its <>type. Not sure, but I would assume so.

moving forward we need to:
1. log into our mssql image bash and
create CommanderApi account (with
correct login creds) and Commander db
with tables and test data.
2. test the api- can we get that test data
from our api/commands endpoint?
what about posting to the db from an endpoint.
This is going to be fun,