We are 99.9% good to go on connecting to the db.
Have a small problem, entity cannot migrate changes (update),
because the user login we are using
in the connection string does not
have the permissions to do anything with
the CommanderDb.

Working on getting this fixed but all
the SQL commands Ive tried fail for
unknown reasons- such as
GRANT ALL ON CommanderDb TO
'CommanderApi'

or

GRANT ALL ON CommanderDb.* TO
'CommanderApi'

Review your INFO605 notes, should be
some viable SQL in the week 7 or 8 ppt.
Once we figured this out, Db problem is solved.