Had a messy ol' time updating to .net core sdk v5.0.0. The old project didnt recieve these changes well. So rather than spending a load of time trying to debug that, I started a new project and transfered the old assets in, redoing the package managment commands in a new csproj file.
Finished this api, done with patch, put and delete.

Since I need to move onto finishing
up older portfolio projects and deploying
things, I decided i will not develop a front
end for this api, and likely not the spring one either. Best case is
we finish up the react thing for Spring
and use swagger ui to auto document this one.